Convert float to double
Convert float to double.
Add DoubleNumerics.
Remove Numerics

+++ PotatoRaytracing/src/PotatoObject.cs
@@ -1,6 +1,6 @@
 ﻿using System.Drawing;
 using System.IO;
-using System.Numerics;
+using System.DoubleNumerics;
 
 namespace PotatoRaytracing
 {
@@ -19,7 +19,7 @@ namespace PotatoRaytracing
 
         public string GetTexturePath() => texturePath;
 
-        public abstract Vector2 GetUV(float x, float y, float z, Bitmap texture);
+        public abstract Vector2 GetUV(double x, double y, double z, Bitmap texture);
         public abstract Vector2 GetUV(Vector3 position, Bitmap texture);
 
         private void LoadTexture(string texturePath)
